+++ math/rabinMiller.cpp
@@ -1,5 +1,5 @@
-ll power(ll x, ll y, ll p) {
-	ll res = 1;
+ll power(int x, int y, int p) {
+ int res = 1;
 	x = x % p;
 	while (y > 0) {
 		if (y & 1) res = (res*x) % p;
@@ -9,9 +9,9 @@ ll power(ll x, ll y, ll p) {
 	return res;
 }
 
-bool miillerTest(ll d, ll n) {
-	ll a = 2 + rand() % (n - 4);
-	ll x = power(a, d, n);
+bool miillerTest(int d, int n) {
+ int a = 2 + rand() % (n - 4);
+ int x = power(a, d, n);
 	if (x == 1  || x == n-1) return true;
 	while (d != n-1) {
 		x = (x * x) % n;
@@ -25,12 +25,12 @@ bool miillerTest(ll d, ll n) {
 // It returns false if n is composite and returns true if n
 // is probably prime.  k is an input parameter that determines
 // accuracy level. Higher value of k indicates more accuracy.
-bool isPrime(ll n, ll k) {
+bool isPrime(int n, int k) {
 	if (n <= 1 || n == 4)  return false;
 	if (n <= 3) return true;
-	ll d = n - 1;
+ int d = n - 1;
 	while (d % 2 == 0) d /= 2;
-	for (ll i = 0; i < k; i++)
+	for (int i = 0; i < k; i++)
 		if (!miillerTest(d, n))
 			return false;
 	return true;
